+++ lsm/remote/__init__.py
@@ -12,6 +12,7 @@ __all__ = [
     # Base classes
     "BaseRemoteProvider",
     "RemoteResult",
+    "BaseOAIProvider",
     # Factory functions
     "create_remote_provider",
     "register_remote_provider",
@@ -58,6 +59,7 @@ from lsm.remote.providers.wikipedia import WikipediaProvider
 from lsm.remote.providers.arxiv import ArXivProvider
 from lsm.remote.providers.semantic_scholar import SemanticScholarProvider
 from lsm.remote.providers.core import COREProvider
+from lsm.remote.providers.base_oai import BaseOAIProvider
 from lsm.remote.providers.oai_pmh import OAIPMHProvider, OAIPMHClient, OAIRecord, KNOWN_REPOSITORIES
 from lsm.remote.providers.philpapers import PhilPapersProvider
 from lsm.remote.providers.ixtheo import IxTheoProvider